perf: Avoid some recursion when unpacking aliases with `AliasRef::typ`
