+++ check/src/unify_type.rs
@@ -117,7 +117,7 @@ impl<T> From<ResolveError> for TypeError<Symbol, T> {
 impl<I, T> fmt::Display for TypeError<I, T>
 where
     I: fmt::Display + AsRef<str>,
-    T: TypeExt<I> + ast::Commented + pos::HasSpan,
+    T: TypeExt<Id = I> + ast::Commented + pos::HasSpan,
 {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         let filter = self.make_filter();
@@ -127,7 +127,7 @@ where
 
 pub fn similarity_filter<'a, I, T>(typ: &'a T, fields: &'a [I]) -> Box<Fn(&I) -> Filter + 'a>
 where
-    T: TypeExt<I>,
+    T: TypeExt<Id = I>,
     I: AsRef<str>,
 {
     let mut field_similarity = typ
@@ -166,7 +166,7 @@ where
 impl<I, T> TypeError<I, T>
 where
     I: fmt::Display + AsRef<str>,
-    T: TypeExt<I> + ast::Commented + pos::HasSpan,
+    T: TypeExt<Id = I> + ast::Commented + pos::HasSpan,
 {
     pub fn make_filter<'a>(&'a self) -> Box<Fn(&I) -> Filter + 'a> {
         match *self {
